+++ test/public/CREATE2/CreateX.deployCreate2AndInit_4Args_CustomiseSalt.t.sol
@@ -55,8 +55,8 @@ contract CreateX_DeployCreate2AndInit_4Args_CustomiseSalt_Public_Test is BaseTes
         whenTheInitCodeSuccessfullyCreatesARuntimeBytecodeWithANonZeroLength
         whenTheInitialisationCallIsSuccessful
     {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, 2 * (values.constructorAmount + values.initCallAmount));
         vm.assume(
             chainId != block.chainid &&
@@ -241,8 +241,8 @@ contract CreateX_DeployCreate2AndInit_4Args_CustomiseSalt_Public_Test is BaseTes
         whenTheCreateXContractHasANonZeroBalance(chainId)
         whenTheRefundTransactionIsSuccessful
     {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, 2 * (values.constructorAmount + values.initCallAmount));
         vm.assume(
             chainId != block.chainid &&
@@ -440,8 +440,8 @@ contract CreateX_DeployCreate2AndInit_4Args_CustomiseSalt_Public_Test is BaseTes
         whenTheCreateXContractHasANonZeroBalance(chainId)
         whenTheRefundTransactionIsUnsuccessful
     {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(SELF, values.constructorAmount + values.initCallAmount);
         vm.assume(chainId != block.chainid && chainId != 0);
         // Helper logic to increase the probability of matching a permissioned deploy protection during fuzzing.
@@ -497,8 +497,8 @@ contract CreateX_DeployCreate2AndInit_4Args_CustomiseSalt_Public_Test is BaseTes
         whenTheInitCodeSuccessfullyCreatesARuntimeBytecodeWithANonZeroLength
         whenTheInitialisationCallIsUnsuccessful
     {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, values.constructorAmount + values.initCallAmount);
         vm.assume(
             chainId != block.chainid &&
@@ -555,8 +555,8 @@ contract CreateX_DeployCreate2AndInit_4Args_CustomiseSalt_Public_Test is BaseTes
         bytes32 salt,
         uint64 chainId
     ) external whenTheInitCodeSuccessfullyCreatesARuntimeBytecodeWithAZeroLength {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, values.constructorAmount + values.initCallAmount);
         vm.assume(
             chainId != block.chainid &&
@@ -609,8 +609,8 @@ contract CreateX_DeployCreate2AndInit_4Args_CustomiseSalt_Public_Test is BaseTes
         bytes32 salt,
         uint64 chainId
     ) external whenTheInitCodeFailsToDeployARuntimeBytecode {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, values.constructorAmount + values.initCallAmount);
         vm.assume(
             chainId != block.chainid &&
