+++ test/public/CREATE3/CreateX.deployCreate3AndInit_3Args.t.sol
@@ -62,8 +62,8 @@ contract CreateX_DeployCreate3AndInit_3Args_Public_Test is BaseTest {
         whenTheInitCodeSuccessfullyCreatesARuntimeBytecodeWithANonZeroLength
         whenTheInitialisationCallIsSuccessful
     {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, 2 * (values.constructorAmount + values.initCallAmount));
         vm.assume(
             chainId != block.chainid &&
@@ -221,8 +221,8 @@ contract CreateX_DeployCreate3AndInit_3Args_Public_Test is BaseTest {
         whenTheCreateXContractHasANonZeroBalance(chainId)
         whenTheRefundTransactionIsSuccessful
     {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, 2 * (values.constructorAmount + values.initCallAmount));
         vm.assume(
             chainId != block.chainid &&
@@ -379,8 +379,8 @@ contract CreateX_DeployCreate3AndInit_3Args_Public_Test is BaseTest {
         whenTheCreateXContractHasANonZeroBalance(amount)
         whenTheRefundTransactionIsUnsuccessful
     {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(SELF, values.constructorAmount + values.initCallAmount);
         vm.startPrank(SELF);
         bytes32 salt = createXHarness.exposed_generateSalt();
@@ -417,8 +417,8 @@ contract CreateX_DeployCreate3AndInit_3Args_Public_Test is BaseTest {
         whenTheInitCodeSuccessfullyCreatesARuntimeBytecodeWithANonZeroLength
         whenTheInitialisationCallIsUnsuccessful
     {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, values.constructorAmount + values.initCallAmount);
         vm.assume(originalDeployer != createXAddr && originalDeployer != zeroAddress);
         vm.startPrank(originalDeployer);
@@ -455,8 +455,8 @@ contract CreateX_DeployCreate3AndInit_3Args_Public_Test is BaseTest {
         address originalDeployer,
         CreateX.Values memory values
     ) external whenTheProxyContractCreationFails {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, values.constructorAmount + values.initCallAmount);
         vm.assume(originalDeployer != createXAddr && originalDeployer != zeroAddress);
         vm.startPrank(originalDeployer);
@@ -493,8 +493,8 @@ contract CreateX_DeployCreate3AndInit_3Args_Public_Test is BaseTest {
         address originalDeployer,
         CreateX.Values memory values
     ) external whenTheInitCodeSuccessfullyCreatesARuntimeBytecodeWithAZeroLength {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, values.constructorAmount + values.initCallAmount);
         vm.assume(originalDeployer != createXAddr && originalDeployer != zeroAddress);
         vm.startPrank(originalDeployer);
@@ -527,8 +527,8 @@ contract CreateX_DeployCreate3AndInit_3Args_Public_Test is BaseTest {
         address originalDeployer,
         CreateX.Values memory values
     ) external whenTheInitCodeFailsToDeployARuntimeBytecode {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, values.constructorAmount + values.initCallAmount);
         vm.assume(originalDeployer != createXAddr && originalDeployer != zeroAddress);
         vm.startPrank(originalDeployer);
