+++ test/public/CREATE3/CreateX.deployCreate3AndInit_5Args.t.sol
@@ -64,8 +64,8 @@ contract CreateX_DeployCreate3AndInit_5Args_Public_Test is BaseTest {
         whenTheInitCodeSuccessfullyCreatesARuntimeBytecodeWithANonZeroLength
         whenTheInitialisationCallIsSuccessful
     {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, 2 * (values.constructorAmount + values.initCallAmount));
         vm.assume(
             chainId != block.chainid &&
@@ -292,8 +292,8 @@ contract CreateX_DeployCreate3AndInit_5Args_Public_Test is BaseTest {
         whenTheCreateXContractHasANonZeroBalance(chainId)
         whenTheRefundTransactionIsSuccessful
     {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, 2 * (values.constructorAmount + values.initCallAmount));
         vm.assume(
             chainId != block.chainid &&
@@ -525,8 +525,8 @@ contract CreateX_DeployCreate3AndInit_5Args_Public_Test is BaseTest {
         whenTheCreateXContractHasANonZeroBalance(chainId)
         whenTheRefundTransactionIsUnsuccessful
     {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(SELF, values.constructorAmount + values.initCallAmount);
         vm.assume(chainId != block.chainid && chainId != 0);
         // Helper logic to increase the probability of matching a permissioned deploy protection during fuzzing.
@@ -584,8 +584,8 @@ contract CreateX_DeployCreate3AndInit_5Args_Public_Test is BaseTest {
         whenTheInitCodeSuccessfullyCreatesARuntimeBytecodeWithANonZeroLength
         whenTheInitialisationCallIsUnsuccessful
     {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, values.constructorAmount + values.initCallAmount);
         vm.assume(
             chainId != block.chainid &&
@@ -644,8 +644,8 @@ contract CreateX_DeployCreate3AndInit_5Args_Public_Test is BaseTest {
         bytes32 salt,
         uint64 chainId
     ) external whenTheProxyContractCreationFails {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, values.constructorAmount + values.initCallAmount);
         vm.assume(
             chainId != block.chainid &&
@@ -704,8 +704,8 @@ contract CreateX_DeployCreate3AndInit_5Args_Public_Test is BaseTest {
         bytes32 salt,
         uint64 chainId
     ) external whenTheInitCodeSuccessfullyCreatesARuntimeBytecodeWithAZeroLength {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, values.constructorAmount + values.initCallAmount);
         vm.assume(
             chainId != block.chainid &&
@@ -760,8 +760,8 @@ contract CreateX_DeployCreate3AndInit_5Args_Public_Test is BaseTest {
         bytes32 salt,
         uint64 chainId
     ) external whenTheInitCodeFailsToDeployARuntimeBytecode {
-        values.constructorAmount = bound(values.constructorAmount, 0, type(uint64).max);
-        values.initCallAmount = bound(values.initCallAmount, 0, type(uint64).max);
+        values.constructorAmount = uint128(bound(values.constructorAmount, 0, type(uint64).max));
+        values.initCallAmount = uint128(bound(values.initCallAmount, 0, type(uint64).max));
         vm.deal(originalDeployer, values.constructorAmount + values.initCallAmount);
         vm.assume(
             chainId != block.chainid &&
